% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotone-spline.R
\name{mspline}
\alias{mspline}
\title{mpline}
\usage{
mspline(x, y, k = 10, lower = NA, upper = NA)
}
\arguments{
\item{x}{predictors}

\item{y}{predicted values}

\item{k}{max number of knots (default 10)}

\item{lower}{optional lower bound}

\item{upper}{optional upper bound}
}
\value{
list of class \code{mspline} containing the fitted elements.
}
\description{
Monotone spline fitting
}
\details{
Monotone spline fitting function using MGCV. Also has a predict method
based on mgcv::Predict.matrix. However, it doesn't really play
according to the rules of the generic.
}
\examples{
# Generate data from a monotonic truth.
x <- runif(100) * 4 - 1
x <- sort(x)
f <- exp(4 * x) / (1 + exp(4 * x))
y <- f + rnorm(100) * 0.1
plot(x, y)
fv <- mspline(x, y, 5)
lines(x, predict(fv, x), col = "red")
fv <- mspline(x, y, 10)
lines(x, predict(fv, x), col = "blue")
legend("bottomright", lty = 1, paste0("k=", c(5, 10)),
       col = c("red", "blue"))
}
\references{
See
\url{https://gist.github.com/willtownes/f598e5c2344043675566603d29b6c2d6}
}
\author{
will townes and gzt
}
